Don't null out the IPC::Connection's XPC connection https://bugs.webkit.org/show_bug.cgi?id=159911 rdar://problem/27018065
Reviewed by Alex Christensen.

The function that nulls out the XPC connection, platformInvalidate(), is called from the connection queue,
whereas the XPC connection is normally accessed from the main thread leading to inconsistencies when the
connection is being invalidated while the main thread is trying to access it.

Fix this by simply never nulling out the XPC connection. It will be released when the IPC::Connection is destroyed anyway.

* Platform/IPC/mac/ConnectionMac.mm:
(IPC::Connection::platformInvalidate):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@203387 268f45cc-cd09-0410-ab3c-d52691b4dbfc