EventTarget.dispatchEvent() parameter should not be nullable https://bugs.webkit.org/show_bug.cgi?id=159897
Reviewed by Benjamin Poulain.

Source/WebCore:

EventTarget.dispatchEvent() parameter should not be nullable:
- https://dom.spec.whatwg.org/#interface-eventtarget

Even though the parameter was marked as nullable in our IDL, our
implementation does a null check and we already throw a TypeError
when calling dispatchEvent(null).

Update our IDL so that it matches the specification and so that
the null check is generated in the bindings instead.

No new tests, rebaseline existing tests.

* dom/EventTarget.cpp:
(WebCore::EventTarget::dispatchEventForBindings):
* dom/EventTarget.h:
* dom/EventTarget.idl:

LayoutTests:

Update layout tests as the message of the TypeError exception being
thrown when calling dispatchEvent(null) is now more helpful.

* fast/dom/Window/dispatchEvent-expected.txt:
* fast/events/dispatchEvent-crash-expected.txt:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@203391 268f45cc-cd09-0410-ab3c-d52691b4dbfc