[Linux] MemoryPressureMonitor fallback code to get memory available in older linux kernels doesn't work https://bugs.webkit.org/show_bug.cgi?id=159970
Reviewed by Antonio Gomes.

We are failing to detect zones when parsing /proc/zoneinfo.

* UIProcess/linux/MemoryPressureMonitor.cpp:
(WebKit::lowWatermarkPages): Use strncmp since we want to know whether the line starts with "Node".
(WebKit::MemoryPressureMonitor::MemoryPressureMonitor): Stop the polling if we fail to get the memory available,
because that means it's not sopported in the system for whatever reason.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@203451 268f45cc-cd09-0410-ab3c-d52691b4dbfc