Fix null handling of HTMLSelectElement.value attribute https://bugs.webkit.org/show_bug.cgi?id=159925
Reviewed by Benjamin Poulain.

Source/WebCore:

Fix null handling of HTMLSelectElement.value attribute:
- https://html.spec.whatwg.org/multipage/forms.html#htmlselectelement

We were treating null as the null String which would end up setting
selectedIndex to -1. However, we should treat null as the String "null"
which would set the selectedIndex to the index of the <option> element
whose value is "null".

Firefox and Chrome match the specification.

Test: fast/dom/HTMLSelectElement/value-null-handling.html

* html/HTMLSelectElement.cpp:
(WebCore::HTMLSelectElement::setValue):
* html/HTMLSelectElement.idl:

LayoutTests:

Add layout test coverage. I have verified that this test is passing in
both Firefox and Chrome.

* fast/dom/HTMLSelectElement/value-null-handling-expected.txt: Added.
* fast/dom/HTMLSelectElement/value-null-handling.html: Added.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@203456 268f45cc-cd09-0410-ab3c-d52691b4dbfc