Remove unnecessary if check from ParkingLot.cpp https://bugs.webkit.org/show_bug.cgi?id=159961
Patch by Rajeev Misra <rajeevmisraforapple@gmail.com> on 2016-07-20
Reviewed by Alex Christensen.

A good practice is to have as less conditional statement
or special cases as possible in code. This change
simply removes a unnecessary "if" statement for
condition which was already evaluated by switch/case
and thus there was no need to evaluate again.

* wtf/ParkingLot.cpp:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@203466 268f45cc-cd09-0410-ab3c-d52691b4dbfc