REGRESSION (r203189): Webpage snapshots are partially or fully blank when edge swiping back https://bugs.webkit.org/show_bug.cgi?id=160000 rdar://problem/27455589
Reviewed by Zalan Bujtas.

r203189 triggered a call to -_updateContentRectsWithState: in the middle of a swipe,
during which we'd compute the visible rect of the view as having zero width, causing us
to not create tiles.

-_navigationGestureDidBegin already snapshots the visibleContentRect in _frozenVisibleContentRect,
but -_visibleContentRect then clipped that with ancestor views, causing the issue.

Fix by having -_visibleContentRect just return _frozenVisibleContentRect if we have one.

* UIProcess/API/Cocoa/WKWebView.mm:
(-[WKWebView _visibleContentRect]):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@203483 268f45cc-cd09-0410-ab3c-d52691b4dbfc