Do not keep invalid IOSurface in ImageBufferData. https://bugs.webkit.org/show_bug.cgi?id=160005 <rdar://problem/27208636>
Reviewed by Simon Fraser.

When we fail to initialize the IOSurface for the accelerated context, we switch over to
the non-accelerated code path. Since ImageBufferData::surface is used to indicate whether
the graphics context is in accelerated mode, we need to reset it when the initialization fails.

Unable to create a test case.

* platform/graphics/cg/ImageBufferCG.cpp:
(WebCore::ImageBuffer::ImageBuffer):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@203514 268f45cc-cd09-0410-ab3c-d52691b4dbfc