[Win] jsc.exe sometimes never exits. https://bugs.webkit.org/show_bug.cgi?id=158073
Reviewed by Mark Lam.

Make sure the VM is deleted after the test has finished. This will gracefully stop the sampling profiler thread,
and give the thread the opportunity to release the machine thread lock aquired in SamplingProfiler::takeSample.
If the sampling profiler thread was terminated while holding the machine thread lock, the machine thread will
not be able to grab the lock afterwards.

* jsc.cpp:
(jscmain):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@203603 268f45cc-cd09-0410-ab3c-d52691b4dbfc