Fullscreen video zoom button does not work after rotating when aspect ratio matches display. https://bugs.webkit.org/show_bug.cgi?id=160263 rdar://problem/27368872
Patch by Jeremy Jones <jeremyj@apple.com> on 2016-07-27
Reviewed by Eric Carlson.

When video and display aspect ratio match, and rotating from landscape to protrait, the transform used in layout
will be Identity. This means checking the transform for identity is an insufficient test to see if the bounds
need to be resolved.

Instead, always attempt to resolve the bounds and do a more accurate test while doing so.

* platform/ios/WebVideoFullscreenInterfaceAVKit.mm:
(-[WebAVPlayerLayer layoutSublayers]):
(-[WebAVPlayerLayer resolveBounds]):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@203792 268f45cc-cd09-0410-ab3c-d52691b4dbfc