Undefined Behavior in JSValue cast from NaN https://bugs.webkit.org/show_bug.cgi?id=160322
Patch by Jonathan Bedard <jbedard@apple.com> on 2016-07-29
Reviewed by Mark Lam.

JSValues can be constructed from doubles, and in some cases, are deliberately constructed with NaN values.

In circumstances where NaN is bound through the default JSValue constructor, however, an undefined conversion
to int32_t occurs.  While the subsequent if statement should fail and construct the JSValue through the explicit
double constructor, given that the deliberate use of NaN is fairly common, it seems that the jsNaN() function
should immediately call the explicit double constructor both for efficiency and to prevent inadvertent
suppressing of any other bugs which may be instantiating a JSValue with a NaN double.

* runtime/JSCJSValueInlines.h:
(JSC::jsNaN): Explicit double construction for NaN JSValues to avoid undefined behavior.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@203925 268f45cc-cd09-0410-ab3c-d52691b4dbfc