Kill [StrictTypeChecking] IDL extended attribute https://bugs.webkit.org/show_bug.cgi?id=160382
Reviewed by Darin Adler.

Kill [StrictTypeChecking] IDL extended attribute. Its only remaining
use was on DOMString parameters for a few CanvasRenderingContext2D
operations in order to work around a bug in our overload resolution.

However, as of r204028, we implement the Web IDL overload resolution
algorithm and [StrictTypeChecking] is now a no-op and is no longer
needed.

* bindings/scripts/CodeGeneratorJS.pm:
(GenerateParametersCheckExpression):
* bindings/scripts/IDLAttributes.txt:
* bindings/scripts/test/TestObj.idl:
* bindings/scripts/test/TestTypedefs.idl:
* html/canvas/CanvasRenderingContext2D.idl:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@204033 268f45cc-cd09-0410-ab3c-d52691b4dbfc