Fix 32-bit OverridesHasInstance in the DFG. https://bugs.webkit.org/show_bug.cgi?id=160600
Reviewed by Mark Lam.

In https://trac.webkit.org/changeset/204140, we fixed an issue where the DFG might
do the wrong thing if it proved that the Symbol.hasInstance value for a constructor
was a constant late in compilation. That fix was ommited from the 32-bit version,
causing the new test to fail.

* dfg/DFGSpeculativeJIT32_64.cpp:
(JSC::DFG::SpeculativeJIT::compile):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@204176 268f45cc-cd09-0410-ab3c-d52691b4dbfc