[Web IDL] Add support for 'any' type in dictionaries https://bugs.webkit.org/show_bug.cgi?id=160511
Reviewed by Sam Weinig.

Add support for 'any' type in dictionaries. It will pass the value as
a JSValue to the implementation. If the member is not present, it will
pass jsUndefined() to the implementation.

Having 'any' support in dictionaries is a convenient workaround for
types we don't support yet (e.g. union types).

No new tests, updated bindings tests.

* bindings/js/JSDOMConvert.h:
(WebCore::Converter<JSC::JSValue>::convert):
* bindings/scripts/test/JS/JSTestObj.cpp:
(WebCore::convert<TestObj::Dictionary>):
* bindings/scripts/test/TestObj.idl:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@204216 268f45cc-cd09-0410-ab3c-d52691b4dbfc