[GTK] Bad performance in accelerated compositing mode with the modesetting intel driver and DRI3 enabled https://bugs.webkit.org/show_bug.cgi?id=160491
Reviewed by Michael Catanzaro.

This is becoming the default in several linux distributions and it makes WebKitGTK+ unusable in accelerated
compositing mode, which is now always enabled because since we switched to use the threaded compositor. The
problem seems to be an optimization of the intel driver for windows that are offscreen, and our redirected
window is always positioned at ScreenWidth + 1, 0. Using negative coordinates to position the redirected window
offscreen fixes the problem.
See also https://bugs.freedesktop.org/show_bug.cgi?id=85064.

* WebProcess/WebPage/gtk/RedirectedXCompositeWindow.cpp:
(WebKit::RedirectedXCompositeWindow::RedirectedXCompositeWindow): Position the parent window at -1, -1.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@204250 268f45cc-cd09-0410-ab3c-d52691b4dbfc