[macOS Sierra] Fix flaky test: media/controls/picture-in-picture.html https://bugs.webkit.org/show_bug.cgi?id=160707
Reviewed by Eric Carlson.

This test became flaky after r201474 when we started to delay showing
the inline placeholder until we are sure the video layer has been moved
into the video fullscreen layer. This means we can't guarantee that the
placeholder is visible right away after the video's presentation mode
changes to "picture-in-picture".

To fix this, we'll update the test so that we'll wait until the placeholder
becomes visible before testing its visibility without the "controls" attribute.

* media/controls/picture-in-picture.html:
* platform/mac-wk2/TestExpectations:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@204304 268f45cc-cd09-0410-ab3c-d52691b4dbfc