[JSC] TinyPtrSet::deleteListIfNecessary() no longer needs to test for reservedValue https://bugs.webkit.org/show_bug.cgi?id=160721
Patch by Benjamin Poulain <bpoulain@apple.com> on 2016-08-10
Reviewed by Sam Weinig.

Previously, TinyPtrSet was using a ThinFlag if there was no OutOfLineList.
A side effect is that isThin() was false for reservedValue. When deleting
the list, we had to consider both cases: isThin() and reservedValue.

In r204065, I changed the flag to a FatFlag to ensure TinyPtrSet can be initialized
with bzero().
This changes the value of isThin() of reservedValue to true since reservedValue
does not have the FatFlag bit set.
This patch removes the useless condition from deleteListIfNecessary().

* wtf/TinyPtrSet.h:
(WTF::TinyPtrSet::deleteListIfNecessary):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@204354 268f45cc-cd09-0410-ab3c-d52691b4dbfc