Gardening: fix gcc builds after r204387.
Not reviewed.

Apparently, gcc is not sophisticated enough to realize that the end of the
function is unreachable, and is wrongly complaining about "control reaches end of
non-void function".  I'm restoring the RELEASE_ASSERT_NOT_REACHED() and return
statement at the end of MarkedBlock::sweepHelper() to appease gcc.

* heap/MarkedBlock.cpp:
(JSC::MarkedBlock::sweepHelper):



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@204398 268f45cc-cd09-0410-ab3c-d52691b4dbfc