Fix compiler errors when building iOS WebKit using the iOS 10 beta SDK https://bugs.webkit.org/show_bug.cgi?id=160725
Reviewed by Sam Weinig.

Source/WebCore:

* platform/cocoa/ThemeCocoa.mm: Unconditionally include header dlfcn.h as it
exists in both the public iOS 9.3 SDK and iOS 10 beta SDK.
* platform/spi/cocoa/CoreTextSPI.h: Add SPI declarations for constants that were
used in r204107.
* platform/spi/cocoa/PassKitSPI.h: Remove unnecessary #import statements when
building with the Apple Internal SDK. Include header PassKit/PassKit.h when
building for iOS.
* platform/spi/cocoa/QuartzCoreSPI.h: No need to define CLayer.contentsFormat
when building with the iOS 10 beta SDK as it is now part of the public API.

Source/WebKit/mac:

OSAtomicCompareAndSwap32() has been deprecated as of the iOS 10 beta SDK. For now,
silence the complier warning.

* WebView/WebView.mm:
(-[WebView _dispatchTileDidDraw:]):

Source/WebKit2:

* Platform/spi/ios/UIKitSPI.h: Add SPI for UITextInputSuggestionDelegate.
* UIProcess/ApplePay/ios/WebPaymentCoordinatorProxyIOS.mm: Remove unnecessary include of PKPaymentMerchantSession.h.

Source/WTF:

For now, disable OS_LOG when building with the iOS 10 beta SDK until
we have the fix for <rdar://problem/27758343>.

* wtf/Platform.h:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@204456 268f45cc-cd09-0410-ab3c-d52691b4dbfc