[Cocoa] Remove dependency on Objective-C bindings in WebKit2 editing code https://bugs.webkit.org/show_bug.cgi?id=160891
Reviewed by Anders Carlsson.

Source/WebCore:

* WebCore.xcodeproj/project.pbxproj: Removed EditorCocoa.h.

* editing/Editor.cpp: Removed member initialization for members now
initialized in the class definition.

* editing/Editor.h: Added FragmentAndResources struct, createFragment
member function, and _WebCreateFragment function. Also initialize all the
scalar data members here instead of in the constructor.

* editing/cocoa/EditorCocoa.h: Removed. This was unused.

* editing/cocoa/EditorCocoa.mm:
(WebCore::Editor::createFragment): Added. Calls through to WebKitLegacy
using soft linking for now, but later should be implemented here in WebCore.

* editing/ios/EditorIOS.mm:
(WebCore::Editor::createFragmentAndAddResources): Call createFragment.

* editing/mac/EditorMac.mm:
(WebCore::Editor::createFragmentAndAddResources): Call createFragment.

* loader/EmptyClients.h: Removed documentFragmentFromAttributedString
implementation.

* page/EditorClient.h: Removed documentFragmentFromAttributedString.
Two reasons: The first is that this is now implemented in WebCore. While
there is a call through to WebKitLegacy, it's a temporary thing and done
with soft linking. The second reason is that this returned a raw pointer,
which is not safe in general. Was safe here because it was returning a
pointer owned by an autoreleased Objective-C object.

Source/WebKit/mac:

* WebCoreSupport/WebEditorClient.h: Removed documentFragmentFromAttributedString.

* WebCoreSupport/WebEditorClient.mm:
(attributesForAttributedStringConversion): Replaced the
createExcludedElementsForAttributedStringConversion function with this one.
No good reason to keep the array around instead of the dictionary.
(_WebCreateFragment): Renamed from documentFragmentFromAttributedString.
This is now a global function exported so it can be called by WebCore with
a different way of returning its value.

* WebKit.exp: Added _WebCreateFragment.

* WebView/WebHTMLView.mm:
(-[WebHTMLView _documentFragmentFromPasteboard:forType:inContext:subresources:]):
Removed long-ago-obsolete workaround for Radar bug 5052369.

Source/WebKit2:

* WebProcess/WebCoreSupport/WebEditorClient.h: Removed documentFragmentFromAttributedString.
* WebProcess/WebCoreSupport/ios/WebEditorClientIOS.mm:
(WebKit::WebEditorClient::documentFragmentFromAttributedString): Deleted.
* WebProcess/WebCoreSupport/mac/WebEditorClientMac.mm:
(WebKit::createExcludedElementsForAttributedStringConversion): Deleted.
(WebKit::WebEditorClient::documentFragmentFromAttributedString): Deleted.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@204508 268f45cc-cd09-0410-ab3c-d52691b4dbfc