Heap::collectAllGarbage() should work with JSC_useImmortalObjects=true. https://bugs.webkit.org/show_bug.cgi?id=160917
Reviewed by Filip Pizlo.

If we do an synchronous GC when JSC_useImmortalObjects=true, we'll get a
RELEASE_ASSERT failure:

    $ JSC_useImmortalObjects=true jsc
    >>> gc()
    Trace/BPT trap: 5

This is because Heap::collectAllGarbage() is doing an explicit sweep of the
MarkedSpace, and the sweeper is expecting to see no RetiredBlocks.  However, we
make objects immortal by retiring their blocks.  As a result, there is a mismatch
in expectancy.

The fix is simply to not run the sweeper when JSC_useImmortalObjects=true.

* heap/Heap.cpp:
(JSC::Heap::collectAllGarbage):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@204542 268f45cc-cd09-0410-ab3c-d52691b4dbfc