Race between creating/deleting a database in test https://bugs.webkit.org/show_bug.cgi?id=161285
Patch by Joseph Pecoraro <pecoraro@apple.com> on 2016-08-27
Reviewed by Darin Adler.

Because IndexedDB creation is asynchronous, we were not waiting for the
database to be completely open before sending more commands. These
could race and give unexpected results. Wait for the database to be
successfully created before interacting with it.

* inspector/indexeddb/resources/utilities.js:
(createEmptyDatabase):
(createDatabaseWithStores):
Send a single when the database creation is completed.

* inspector/indexeddb/clearObjectStore-expected.txt:
* inspector/indexeddb/clearObjectStore.html:
* inspector/indexeddb/deleteDatabaseNamesWithSpace.html:
* inspector/indexeddb/requestData-expected.txt:
* inspector/indexeddb/requestDatabase-expected.txt:
* inspector/indexeddb/requestDatabase.html:
* inspector/indexeddb/requestDatabaseNames.html:
Listen for the database created signal before proceeding
with the test.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@205086 268f45cc-cd09-0410-ab3c-d52691b4dbfc