[JSC] Some arith nodes are too pessimistic with the types supported on the fast path https://bugs.webkit.org/show_bug.cgi?id=161410
Patch by Benjamin Poulain <bpoulain@apple.com> on 2016-08-30
Reviewed by Geoffrey Garen.

* dfg/DFGFixupPhase.cpp:
(JSC::DFG::FixupPhase::fixupNode):
DoubleRep is able to convert numbers, undefined, booleans and null.
I was too pessimistic when I gated the double implementations
on number-or-boolean speculation. We can just let DoubleRep convert
the other cases as long as it is not a Cell.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@205216 268f45cc-cd09-0410-ab3c-d52691b4dbfc